feat: url sync
